[SFU] Split sfu-courses.py into course.py and api.py - Change line endings to UNIX. - Change course command to use new discord.py API methods. - Still need to clean up api.py.
